Added checks before exporting project folder
During export of the project folder, if 3D print check- Check geometry is marked, the code will check for non-manifolds, intersections and sharp edges in the imported graded meshes. If there are any problems found, the project is not exported and the errors are reported in a text file called "Error_Checkall_info.txt" which is saved to the directory.